searchタグを追加。それに伴いtextタグとテストコードをリファクタリング。
