+++ src/main/java/nablarch/common/web/tag/SearchTag.java
@@ -1,0 +1,12 @@
+package nablarch.common.web.tag;
+
+/**
+ * 入力データ復元とHTMLエスケープを行うinputタグ(type="search")を出力するクラス。
+ * @author Kiyohito Itoh
+ */
+public class SearchTag extends TextTag {
+    @Override
+    protected String getTagName() {
+        return "search";
+    }
+}
