+++ src/test/java/nablarch/common/web/tag/SearchTagTest.java
@@ -1,0 +1,10 @@
+package nablarch.common.web.tag;
+
+/**
+ * @author Kiyohito Itoh
+ */
+public class SearchTagTest extends TextTagTestSupport<SearchTag> {
+    public SearchTagTest() {
+        super(new SearchTag(), "search");
+    }
+}
